Move skip sx test from silx.sx.test to silx.test
Conflicts:
	silx/sx/test/test_sx.py
